[CHORE] Move component in separate folders
