+++ src/pages/Auth/SignUp.js
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { Link } from "react-router-dom";
-import { useAuth } from "../context/auth-context";
+import { useAuth } from "../../context/auth-context";
 import "./Auth.css";
 
 export const SignUp = () => {
