+++ src/pages/index.js
@@ -1,9 +1,9 @@
-export * from "./Home";
-export * from "./ProductPage";
-export * from "./Login";
-export * from "./SignUp";
-export * from "./Cart";
-export * from "./Wishlist";
-export * from "./PageNotFound";
+export * from "./Home/Home";
+export * from "./Products/ProductPage";
+export * from "./Auth/Login";
+export * from "./Auth/SignUp";
+export * from "./Cart/Cart";
+export * from "./Wishlist/Wishlist";
+export * from "./PageNotFound/PageNotFound";
 export * from "./OrderSummary/OrderSummary";
 export * from "./ThankYou/ThankYou";
