+++ src/shared/index.js
@@ -1,2 +1,2 @@
 export * from './Navbar/Navbar';
-export * from './Footer';
+export * from './Footer/Footer';
